+++ Source/MediaStorageAndFileFormat/gdcmSegmentWriter.cxx
@@ -27,7 +27,7 @@ SegmentWriter::~SegmentWriter()
 
 unsigned int SegmentWriter::GetNumberOfSegments() const
 {
-  return Segments.size();
+  return (unsigned int)Segments.size();
 }
 
 void SegmentWriter::SetNumberOfSegments(const unsigned int size)
@@ -83,11 +83,11 @@ bool SegmentWriter::PrepareWrite()
   // Fill the Segment Sequence
   const unsigned int              numberOfSegments  = this->GetNumberOfSegments();
   assert( numberOfSegments );
-  const unsigned int              nbItems           = segmentsSQ->GetNumberOfItems();
+  const size_t nbItems           = segmentsSQ->GetNumberOfItems();
   if (nbItems < numberOfSegments)
   {
-    const int          diff           = numberOfSegments - nbItems;
-    const unsigned int nbOfItemToMake = (diff > 0?diff:0);
+    const size_t diff           = numberOfSegments - nbItems;
+    const size_t nbOfItemToMake = (diff > 0?diff:0);
     for(unsigned int i = 1; i <= nbOfItemToMake; ++i)
     {
       Item item;
@@ -113,7 +113,7 @@ bool SegmentWriter::PrepareWrite()
     Attribute<0x0062, 0x0004> segmentNumberAt;
     unsigned short segmentNumber = segment->GetSegmentNumber();
     if (segmentNumber == 0)
-      segmentNumber = itemNumber;
+      segmentNumber = (unsigned short)itemNumber;
     segmentNumberAt.SetValue( segmentNumber );
     segmentDS.Replace( segmentNumberAt.GetAsDataElement() );
 
@@ -176,7 +176,7 @@ bool SegmentWriter::PrepareWrite()
       anatRegSQ->SetLengthToUndefined();
 
       // Fill the Anatomic Region Sequence
-      const unsigned int nbItems = anatRegSQ->GetNumberOfItems();
+      const size_t nbItems = anatRegSQ->GetNumberOfItems();
       if (nbItems < 1)  // Only one item is a type 1
       {
         Item item;
@@ -230,7 +230,7 @@ bool SegmentWriter::PrepareWrite()
       propCatSQ->SetLengthToUndefined();
 
       // Fill the Segmented Property Category Code Sequence
-      const unsigned int nbItems = propCatSQ->GetNumberOfItems();
+      const size_t nbItems = propCatSQ->GetNumberOfItems();
       if (nbItems < 1)  // Only one item is a type 1
       {
         Item item;
@@ -284,7 +284,7 @@ bool SegmentWriter::PrepareWrite()
       propTypeSQ->SetLengthToUndefined();
 
       // Fill the Segmented Property Type Code Sequence
-      const unsigned int nbItems = propTypeSQ->GetNumberOfItems();
+      const size_t nbItems = propTypeSQ->GetNumberOfItems();
       if (nbItems < 1)  // Only one item is a type 1
       {
         Item item;
@@ -320,7 +320,7 @@ bool SegmentWriter::PrepareWrite()
     {
       // Surface Count
       Attribute<0x0066, 0x002A> surfaceCountAt;
-      surfaceCountAt.SetValue( surfaceCount );
+      surfaceCountAt.SetValue( (unsigned int)surfaceCount );
       segmentDS.Replace( surfaceCountAt.GetAsDataElement() );
 
       //*****   Referenced Surface Sequence   *****//
@@ -338,11 +338,11 @@ bool SegmentWriter::PrepareWrite()
       segmentsRefSQ->SetLengthToUndefined();
 
       // Fill the Segment Surface Generation Algorithm Identification Sequence
-      const unsigned int   nbItems        = segmentsRefSQ->GetNumberOfItems();
+      const size_t nbItems        = segmentsRefSQ->GetNumberOfItems();
       if (nbItems < surfaceCount)
       {
-        const int          diff           = surfaceCount - nbItems;
-        const unsigned int nbOfItemToMake = (diff > 0?diff:0);
+        const size_t diff           = surfaceCount - nbItems;
+        const size_t nbOfItemToMake = (diff > 0?diff:0);
         for(unsigned int i = 1; i <= nbOfItemToMake; ++i)
         {
           Item item;
@@ -371,7 +371,7 @@ bool SegmentWriter::PrepareWrite()
           refSurfaceNumber = surfaceNumber++;
           surface->SetSurfaceNumber( refSurfaceNumber );
         }
-        refSurfaceNumberAt.SetValue( refSurfaceNumber );
+        refSurfaceNumberAt.SetValue( (unsigned int)refSurfaceNumber );
         segmentsRefDS.Replace( refSurfaceNumberAt.GetAsDataElement() );
 
         //*****   Segment Surface Source Instance Sequence   *****//
